Add Confirm Popup for deleting

+++ src/views/JobView.vue
@@ -17,9 +17,12 @@ const state = reactive({
 
 const deleteJob = async () => {
   try{
-    await axios.delete(`/api/jobs/${jobId}`);
-    toast.success('Job Deleted Successfully');
-    router.push('/jobs');
+    const confirm = window.confirm('Are you sure to delete this job?');
+    if(confirm){
+      await axios.delete(`/api/jobs/${jobId}`);
+      toast.success('Job Deleted Successfully');
+      router.push('/jobs');
+    }
   }catch(error){
     console.error('Error deleting Job', error);
     toast.error('Job Not deleted');
